Update API for HW and SoftwareSerial interface
Getting `SoftwareSerial` to work across multiple platforms is a pain. And
atempting to abstract everything from the user turned out to be a source
of many problems. Such as #21 #41 #61 and #65 to name a few...

As a result, the library will move away from handling its own instance
of `SoftwareSerial` and instead utilize the `Stream` interface. During
instantiation, user will pass the SWSerial object as a reference which
will be cased to a Stream.

The HW inteface is also updated to utilize references instead of
pointers in order to make the API more streamlined. In the future
passing pointers will be completely deprecated.