Fix overflow bug at rate limit check
During the Rate limit check, the milis() function is compared against
the sum of _lastRead and UPDATE_TIME. This can result in a overflow bug.
If the sum happens to be just under the overlow but the millis fuction
does overflow, it will take really long time before the condition passes
again.

The UPDATE_TIME must be subtracted from the millis() function!